get id from local storage then pass the id to logout route, if response status equal to 200, clear the local storage and navigate user to login page

+++ client/src/components/Logout.jsx
@@ -3,12 +3,17 @@ import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import axios from "axios";
 import { BiPowerOff } from "react-icons/bi";
+import { logoutRoute } from "../utils/APIRoutes";
 
 function Logout() {
   const navigate = useNavigate();
   const handleClick = async () => {
-    localStorage.clear();
-    navigate("/login");
+    const id = await JSON.parse(localStorage.getItem("chat-app-user"))._id;
+    const data = await axios.get(`${logoutRoute}/${id}`);
+    if (data.status === 200) {
+      localStorage.clear();
+      navigate("/login");
+    }
   };
   return (
     <Button onClick={handleClick}>
